backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
supabase==2.7.0
PyJWT==2.9.0
google-generativeai==0.8.0
anthropic==0.34.0
reportlab==4.2.0
Pillow==10.4.0
